@@ -8,6 +8,7 @@ import com.aonufrei.pathfindingapp.pathfinding.model.Point2d;
 import com.aonufrei.pathfindingapp.service.rest.PathfindingRestService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -19,13 +20,10 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("api/v1/find-path")
 @Tag(name = "API")
 @CrossOrigin(origins = "*")
+@AllArgsConstructor
 public class PathfindingRestController {
 
 	private final PathfindingRestService pathfindingRestService;
-
-	public PathfindingRestController(PathfindingRestService pathfindingRestService) {
-		this.pathfindingRestService = pathfindingRestService;
-	}
 
 	@PostMapping("grid")
 	@Operation(summary = "Find shortest path on grid map", description = "Finds shortest path between two point on a grid-like map.")
